Refactor marketplace post creation into separate component
